services/api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-pubsub
google-cloud-firestore
python-multipart
python-magic
PyPDF2
pillow
